refactor: render-helpers
