--- packages/mp-loader/demo/index.js
@@ -1,1 +1,0 @@
-import '../!./app.js';
