+++ packages/mp-loader/package.json
@@ -63,6 +63,7 @@
     "sinon": "^6.0.1",
     "sinon-chai": "^3.2.0",
     "webpack": "^4.14.0",
+    "webpack-cli": "^3.1.2",
     "webpack-dev-server": "^3.1.10",
     "webpack-sources": "^1.3.0"
   }
