+++ packages/mp-loader/src/paths.js
@@ -4,5 +4,5 @@
 module.exports = {
   createApp: require.resolve('mp-runtime/src/createApp.js'),
   createPage: require.resolve('mp-runtime/src/createPage.js'),
-  vdomHelper: require.resolve('sfc-runtime/src/vdom'),
+  vdomHelper: require.resolve('render-helpers'),
 };
