+++ packages/render-helpers/README.md
@@ -1,0 +1,7 @@
+# render-helpers
+> Render helpers for mp and sfc loader
+
+```js
+import renderHelpers from 'render-helpers';
+renderHelpers._r(Rax) // renderHelpers._r(React)
+```
