+++ packages/render-helpers/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "render-helpers",
+  "version": "0.0.1",
+  "description": "Render helpers for mp and sfc loader",
+  "main": "lib/index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [
+    "helpers"
+  ],
+  "author": "",
+  "license": "ISC"
+}
