+++ packages/render-helpers/src/createElement.js
@@ -1,8 +1,27 @@
-import { processPropsData } from '../utils';
-import Host from '../host';
+import Host from './host';
+
+/**
+* support for include template
+*/
+function processPropsData(props = {}, self = {}) {
+  const data = Object.create(props.data || {});
+  const { pageInstance } = props;
+
+  Object.keys(props.data || {})
+    .forEach((key) => {
+      const val = props.data[key];
+      if (typeof val === 'string') {
+        data[key] =
+          pageInstance && typeof pageInstance[val] === 'function'
+            ? pageInstance[val].bind(pageInstance)
+            : val;
+      }
+    });
+  return data;
+}
 
 export default function(type, props, children) {
-  const createElement = Host.createElement;
+  const createElement = Host.render.createElement;
 
   if (typeof type === 'object' && type.__esModule === true) {
     type = type.default;
