+++ packages/render-helpers/src/createEmptyElement.js
@@ -1,0 +1,3 @@
+export default function(text) {
+  return null;
+};
