+++ packages/render-helpers/src/setRender.js
@@ -1,0 +1,5 @@
+import Host from './host';
+
+export default function setRender(render) {
+  Host.render = render;
+}
