--- packages/sfc-runtime/src/host.js
@@ -1,6 +1,0 @@
-/**
- * Host share state in runtime
- */
-export default {
-  createElement: null,
-};
