--- packages/sfc-runtime/src/vdom/createEmptyElement.js
@@ -1,3 +1,0 @@
-module.exports = function(text) {
-  return null;
-};
