--- packages/sfc-runtime/src/vdom/setRuntime.js
@@ -1,5 +1,0 @@
-import Host from '../host';
-
-export default function setRuntime({ createElement }) {
-  Host.createElement = createElement;
-}
